[Bronze II] Title: Hashing, Time: 0 ms, Memory: 2020 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.cc
@@ -11,12 +11,12 @@ int main(){
     int l;
     cin >> l;
 
-    int h = 0;
+    long long h = 0ll;
     for(int i = 0; i < l; i++){
         char c;
         cin >> c;
 
-        int tmp = c - 'a' + 1;
+        long long tmp = c - 'a' + 1;
         for(int j = 0; j < i; j++) tmp = (tmp * r) % m;
 
         h = (h + tmp) % m;
